#26: Additions to download script
