+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__
 output
 *.geojson
 *.png
+config.json
